cleanup: cargo clippy --fix 修复 70+ lint + 加 clippy::all gate
needless_range_loop/manual_memcpy/match_like_matches_macro/clamp/
vec_init_then_push/explicit_counter_loop/type_complexity/
too_many_arguments 等全量自动+手动修复。3 个 pipeline 测试预期同步
Speaking(true) 首帧事件（pre-existing 未同步）。

fixes 共性4(clippy)

+++ crates/desktop/src/result_window.rs
@@ -142,7 +142,7 @@ pub fn start_click_through_poller(app: tauri::AppHandle) {
             // 统一用物理坐标：cursor_position() 和 outer_position() 都是 PhysicalPosition，
             // 避免多屏不同缩放率下逻辑/物理混合换算错误
             let (mx, my) = match win.cursor_position() {
-                Ok(p) => (p.x as f64, p.y as f64),
+                Ok(p) => (p.x, p.y),
                 Err(_) => continue,
             };
             let (wx, wy) = match win.outer_position() {
